+++ CHANGELOG.md
@@ -1,10 +1,19 @@
 # Change Log
 
+## Version 1.1.0
+Date: 2015-04-02
+
+**Enhancements:**
+- Support selecting a dimension by clicking on the section in the funnel chart.
+
 ## Version 1.0.1
 Date: 2015-04-02
 
 *Fixes:**
 - Fixed an error when less than 3 dimensions are available
+
+**Enhancements:**
+- Update to the newest version of d3-funnel (v0.6.6)
 
 
 ## Version 1.0.0